Added more Enemies!!
Created child scenes of the Enemy Parent Scene and adjusted each one to create a new Enemy. Created a 'LinkedIn Warrior" enemy, a "GitHub" enemy, and also added the "UTSA Pisser" as well. (Could be a mini boss or something idk) Used chatgpt for the "sprites" just as placeholders for now, feel free to change if you want.